requirements.txt
aiohttp==3.1.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==19.1.0
botbuilder-core==4.0.0a6
botbuilder-schema==4.0.0a6
botframework-connector==4.0.0a6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
idna==2.8
isodate==0.6.0
msrest==0.4.29
multidict==4.5.2
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.5.3
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.3
yarl==1.3.0